Applied password patch
By applying this patch, dmenu will not directly display the keyboard
input, but instead replace it with dots. All data from stdin will be
ignored.

The patch has been changed a bit to have '-' as the character showed.
Also fixed the patch as the diff was done the wrong way.